--- python/test/c_cpp/__init__.py
@@ -1,3 +1,0 @@
-from .factories import Factories
-
-__all__ = ['Factories']
