+++ python/test/c_cpp/test_ast_finder.py
@@ -3,8 +3,8 @@ from unittest import TestCase
 from parameterized import parameterized
 from syntax_tree import ASTFinder, ASTNode
 
-from test.c_cpp.factories import Factories
-from test.clang.test_model_loader import TestModelLoader
+from .factories import Factories
+from test.syntax_tree.model_loader import ModelLoader
 
 class TestFinder(TestCase):
     pass
@@ -13,7 +13,7 @@ class TestKindFinder(TestFinder):
 
     @parameterized.expand(Factories.factories)
     def test_find_bogus(self, _, factory):
-        model = TestModelLoader.load_model(factory)
+        model = ModelLoader.load_model(factory)
         iter = ASTFinder.find_kind(model, '(?i).*bogus.*')
         total = len(list(iter))
         self.assertEqual( total, 0)
@@ -21,7 +21,7 @@ class TestKindFinder(TestFinder):
 
     @parameterized.expand(Factories.factories)
     def test_find_expr(self, _, factory):
-        model = TestModelLoader.load_model(factory)
+        model = ModelLoader.load_model(factory)
         iter = ASTFinder.find_kind(model, '(?i).*expr.*')
         total = len(list(iter))
         self.assertGreater( total, 0)
@@ -31,7 +31,7 @@ class TestAllFinder(TestFinder):
 
     @parameterized.expand(Factories.factories)
     def test_find_all_bogus(self, _, factory):
-        model = TestModelLoader.load_model(factory)
+        model = ModelLoader.load_model(factory)
         def isBogus(node: ASTNode):
             if 'Bogus' in node.get_kind(): yield node
         iter = ASTFinder.find_all(model, isBogus)
@@ -41,7 +41,7 @@ class TestAllFinder(TestFinder):
 
     @parameterized.expand(Factories.factories)
     def test_find_all_expr(self, _, factory):
-        model = TestModelLoader.load_model(factory)
+        model = ModelLoader.load_model(factory)
         def isBinaryOperator(node: ASTNode):
             if 'BINARY_OPERATOR' in node.get_kind(): yield node
         iter = ASTFinder.find_all(model, isBinaryOperator)
