+++ python/test/syntax_tree/model_loader.py
@@ -1,8 +1,9 @@
 from pathlib import Path
 from syntax_tree.ast_factory import ASTFactory
 
-class TestModelLoader():
+class ModelLoader():
 
     @staticmethod
     def load_model(factory:ASTFactory):
+        # note: make sure to load a corresponding model for the language
         return  factory.create(Path(__file__).parent.parent.parent.parent / 'c/src/main.c')
